# backend/pkg/api/marketplace_handlers.go
package api

import (
	"encoding/json"
	"log"
	"net/http"
	"strconv"

	"github.com/go-chi/chi/v5"
	"github.com/google/uuid"
	"github.com/digi604/swarmmarket/backend/internal/common"
	"github.com/digi604/swarmmarket/backend/internal/marketplace"
	"github.com/digi604/swarmmarket/backend/pkg/middleware"
)

// MarketplaceHandler handles marketplace HTTP requests.
type MarketplaceHandler struct {
	service *marketplace.Service
}

// NewMarketplaceHandler creates a new marketplace handler.
func NewMarketplaceHandler(service *marketplace.Service) *MarketplaceHandler {
	return &MarketplaceHandler{service: service}
}

// --- Listings ---

// CreateListing handles creating a new listing.
func (h *MarketplaceHandler) CreateListing(w http.ResponseWriter, r *http.Request) {
	agent := middleware.GetAgent(r.Context())
	if agent == nil {
		common.WriteError(w, http.StatusUnauthorized, common.ErrUnauthorized("not authenticated"))
		return
	}

	var req marketplace.CreateListingRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		common.WriteError(w, http.StatusBadRequest, common.ErrBadRequest("invalid request body"))
		return
	}

	listing, err := h.service.CreateListing(r.Context(), agent.ID, &req)
	if err != nil {
		common.WriteError(w, http.StatusBadRequest, common.ErrBadRequest(err.Error()))
		return
	}

	common.WriteJSON(w, http.StatusCreated, listing)
}

// GetListing handles getting a listing by ID or slug.
func (h *MarketplaceHandler) GetListing(w http.ResponseWriter, r *http.Request) {
	idOrSlug := chi.URLParam(r, "id")

	var listing *marketplace.Listing
	var err error

	// Try parsing as UUID first
	if id, parseErr := uuid.Parse(idOrSlug); parseErr == nil {
		listing, err = h.service.GetListing(r.Context(), id)
	} else {
		// Fall back to slug lookup
		listing, err = h.service.GetListingBySlug(r.Context(), idOrSlug)
	}

	if err != nil {
		if err == marketplace.ErrListingNotFound {
			common.WriteError(w, http.StatusNotFound, common.ErrNotFound("listing not found"))
			return
		}
		log.Printf("[ERROR] GetListing failed: %v (idOrSlug=%s)", err, idOrSlug)
		common.WriteError(w, http.StatusInternalServerError, common.ErrInternalServer("failed to get listing"))
		return
	}

	common.WriteJSON(w, http.StatusOK, listing)
}

// SearchListings handles searching for listings.
func (h *MarketplaceHandler) SearchListings(w http.ResponseWriter, r *http.Request) {
	params := marketplace.SearchListingsParams{
		Query:  r.URL.Query().Get("q"),
		Limit:  parseIntParam(r, "limit", 20),
		Offset: parseIntParam(r, "offset", 0),
	}

	if typeStr := r.URL.Query().Get("type"); typeStr != "" {
		t := marketplace.ListingType(typeStr)
		params.ListingType = &t
	}
	if scopeStr := r.URL.Query().Get("scope"); scopeStr != "" {
		s := marketplace.GeographicScope(scopeStr)
		params.GeographicScope = &s
	}
	if categoryStr := r.URL.Query().Get("category"); categoryStr != "" {
		if catID, err := uuid.Parse(categoryStr); err == nil {
			params.CategoryID = &catID
		}
	}
	if minStr := r.URL.Query().Get("min_price"); minStr != "" {
		if min, err := strconv.ParseFloat(minStr, 64); err == nil {
			params.MinPrice = &min
		}
	}
	if maxStr := r.URL.Query().Get("max_price"); maxStr != "" {
		if max, err := strconv.ParseFloat(maxStr, 64); err == nil {
			params.MaxPrice = &max
		}
	}

	result, err := h.service.SearchListings(r.Context(), params)
	if err != nil {
		log.Printf("[ERROR] SearchListings failed: %v (query=%q, limit=%d, offset=%d)", err, params.Query, params.Limit, params.Offset)
		common.WriteError(w, http.StatusInternalServerError, common.ErrInternalServer("failed to search listings"))
		return
	}

	common.WriteJSON(w, http.StatusOK, result)
}

// DeleteListing handles deleting a listing.
func (h *MarketplaceHandler) DeleteListing(w http.ResponseWriter, r *http.Request) {
	agent := middleware.GetAgent(r.Context())
	if agent == nil {
		common.WriteError(w, http.StatusUnauthorized, common.ErrUnauthorized("not authenticated"))
		return
	}

	idStr := chi.URLParam(r, "id")
	id, err := uuid.Parse(idStr)
	if err != nil {
		common.WriteError(w, http.StatusBadRequest, common.ErrBadRequest("invalid listing id"))
		return
	}

	if err := h.service.DeleteListing(r.Context(), id, agent.ID); err != nil {
		if err == marketplace.ErrListingNotFound {
			common.WriteError(w, http.StatusNotFound, common.ErrNotFound("listing not found"))
			return
		}
		log.Printf("[ERROR] DeleteListing failed: %v (id=%s, agentID=%s)", err, id, agent.ID)
		common.WriteError(w, http.StatusInternalServerError, common.ErrInternalServer("failed to delete listing"))
		return
	}

	w.WriteHeader(http.StatusNoContent)
}

// PurchaseListing handles purchasing a listing.
func (h *MarketplaceHandler) PurchaseListing(w http.ResponseWriter, r *http.Request) {
	agent := middleware.GetAgent(r.Context())
	if agent == nil {
		common.WriteError(w, http.StatusUnauthorized, common.ErrUnauthorized("not authenticated"))
		return
	}

	listingIDStr := chi.URLParam(r, "id")
	listingID, err := uuid.Parse(listingIDStr)
	if err != nil {
		common.WriteError(w, http.StatusBadRequest, common.ErrBadRequest("invalid listing id"))
		return
	}

	var req marketplace.PurchaseListingRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		// Default to quantity of 1 if no body provided
		req.Quantity = 1
	}
	if req.Quantity <= 0 {
		req.Quantity = 1
	}

	result, err := h.service.PurchaseListing(r.Context(), agent.ID, listingID, req.Quantity)
	if err != nil {
		if err == marketplace.ErrListingNotFound {
			common.WriteError(w, http.StatusNotFound, common.ErrNotFound("listing not found"))
			return
		}
		log.Printf("[ERROR] PurchaseListing failed: %v (listingID=%s, agentID=%s)", err, listingID, agent.ID)
		common.WriteError(w, http.StatusBadRequest, common.ErrBadRequest(err.Error()))
		return
	}

	common.WriteJSON(w, http.StatusCreated, result)
}

// --- Requests ---

// CreateRequest handles creating a new request.
func (h *MarketplaceHandler) CreateRequest(w http.ResponseWriter, r *http.Request) {
	agent := middleware.GetAgent(r.Context())
	if agent == nil {
		common.WriteError(w, http.StatusUnauthorized, common.ErrUnauthorized("not authenticated"))
		return
	}

	var req marketplace.CreateRequestRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		common.WriteError(w, http.StatusBadRequest, common.ErrBadRequest("invalid request body"))
		return
	}

	request, err := h.service.CreateRequest(r.Context(), agent.ID, &req)
	if err != nil {
		common.WriteError(w, http.StatusBadRequest, common.ErrBadRequest(err.Error()))
		return
	}

	common.WriteJSON(w, http.StatusCreated, request)
}

// GetRequest handles getting a request by ID.
func (h *MarketplaceHandler) GetRequest(w http.ResponseWriter, r *http.Request) {
	idOrSlug := chi.URLParam(r, "id")

	var request *marketplace.Request
	var err error

	// Try parsing as UUID first
	if id, parseErr := uuid.Parse(idOrSlug); parseErr == nil {
		request, err = h.service.GetRequest(r.Context(), id)
	} else {
		// Fall back to slug lookup
		request, err = h.service.GetRequestBySlug(r.Context(), idOrSlug)
	}

	if err != nil {
		if err == marketplace.ErrRequestNotFound {
			common.WriteError(w, http.StatusNotFound, common.ErrNotFound("request not found"))
			return
		}
		log.Printf("[ERROR] GetRequest failed: %v (idOrSlug=%s)", err, idOrSlug)
		common.WriteError(w, http.StatusInternalServerError, common.ErrInternalServer("failed to get request"))
		return
	}

	common.WriteJSON(w, http.StatusOK, request)
}

// UpdateRequest handles updating a request.
func (h *MarketplaceHandler) UpdateRequest(w http.ResponseWriter, r *http.Request) {
	agent := middleware.GetAgent(r.Context())
	if agent == nil {
		common.WriteError(w, http.StatusUnauthorized, common.ErrUnauthorized("not authenticated"))
		return
	}

	idStr := chi.URLParam(r, "id")
	id, err := uuid.Parse(idStr)
	if err != nil {
		common.WriteError(w, http.StatusBadRequest, common.ErrBadRequest("invalid request id"))
		return
	}

	var req marketplace.UpdateRequestRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		common.WriteError(w, http.StatusBadRequest, common.ErrBadRequest("invalid request body"))
		return
	}

	request, err := h.service.UpdateRequest(r.Context(), agent.ID, id, &req)
	if err != nil {
		if err == marketplace.ErrRequestNotFound {
			common.WriteError(w, http.StatusNotFound, common.ErrNotFound("request not found"))
			return
		}
		log.Printf("[ERROR] UpdateRequest failed: %v (id=%s, agentID=%s)", err, id, agent.ID)
		common.WriteError(w, http.StatusBadRequest, common.ErrBadRequest(err.Error()))
		return
	}

	common.WriteJSON(w, http.StatusOK, request)
}

// SearchRequests handles searching for requests.
func (h *MarketplaceHandler) SearchRequests(w http.ResponseWriter, r *http.Request) {
	params := marketplace.SearchRequestsParams{
		Query:  r.URL.Query().Get("q"),
		SortBy: r.URL.Query().Get("sort"),
		Limit:  parseIntParam(r, "limit", 20),
		Offset: parseIntParam(r, "offset", 0),
	}

	// Default sort is by budget (highest first)
	if params.SortBy == "" {
		params.SortBy = "budget_high"
	}

	if typeStr := r.URL.Query().Get("type"); typeStr != "" {
		t := marketplace.ListingType(typeStr)
		params.RequestType = &t
	}
	if scopeStr := r.URL.Query().Get("scope"); scopeStr != "" {
		s := marketplace.GeographicScope(scopeStr)
		params.GeographicScope = &s
	}

	result, err := h.service.SearchRequests(r.Context(), params)
	if err != nil {
		log.Printf("[ERROR] SearchRequests failed: %v (query=%q, limit=%d, offset=%d)", err, params.Query, params.Limit, params.Offset)
		common.WriteError(w, http.StatusInternalServerError, common.ErrInternalServer("failed to search requests"))
		return
	}

	common.WriteJSON(w, http.StatusOK, result)
}

// --- Offers ---

// SubmitOffer handles submitting an offer to a request.
func (h *MarketplaceHandler) SubmitOffer(w http.ResponseWriter, r *http.Request) {
	agent := middleware.GetAgent(r.Context())
	if agent == nil {
		common.WriteError(w, http.StatusUnauthorized, common.ErrUnauthorized("not authenticated"))
		return
	}

	requestIDStr := chi.URLParam(r, "id")
	requestID, err := uuid.Parse(requestIDStr)
	if err != nil {
		common.WriteError(w, http.StatusBadRequest, common.ErrBadRequest("invalid request id"))
		return
	}

	var req marketplace.CreateOfferRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		common.WriteError(w, http.StatusBadRequest, common.ErrBadRequest("invalid request body"))
		return
	}

	offer, err := h.service.SubmitOffer(r.Context(), agent.ID, requestID, &req)
	if err != nil {
		common.WriteError(w, http.StatusBadRequest, common.ErrBadRequest(err.Error()))
		return
	}

	common.WriteJSON(w, http.StatusCreated, offer)
}

// GetOffers handles getting all offers for a request.
func (h *MarketplaceHandler) GetOffers(w http.ResponseWriter, r *http.Request) {
	requestIDStr := chi.URLParam(r, "id")
	requestID, err := uuid.Parse(requestIDStr)
	if err != nil {
		common.WriteError(w, http.StatusBadRequest, common.ErrBadRequest("invalid request id"))
		return
	}

	offers, err := h.service.GetOffersByRequest(r.Context(), requestID)
	if err != nil {
		log.Printf("[ERROR] GetOffers failed: %v (requestID=%s)", err, requestID)
		common.WriteError(w, http.StatusInternalServerError, common.ErrInternalServer("failed to get offers"))
		return
	}

	common.WriteJSON(w, http.StatusOK, map[string]any{"offers": offers})
}

// AcceptOffer handles accepting an offer.
func (h *MarketplaceHandler) AcceptOffer(w http.ResponseWriter, r *http.Request) {
	agent := middleware.GetAgent(r.Context())
	if agent == nil {
		common.WriteError(w, http.StatusUnauthorized, common.ErrUnauthorized("not authenticated"))
		return
	}

	offerIDStr := chi.URLParam(r, "offerId")
	offerID, err := uuid.Parse(offerIDStr)
	if err != nil {
		common.WriteError(w, http.StatusBadRequest, common.ErrBadRequest("invalid offer id"))
		return
	}

	offer, err := h.service.AcceptOffer(r.Context(), agent.ID, offerID)
	if err != nil {
		common.WriteError(w, http.StatusBadRequest, common.ErrBadRequest(err.Error()))
		return
	}

	common.WriteJSON(w, http.StatusOK, offer)
}

// --- Categories ---

// GetCategories handles getting all categories.
func (h *MarketplaceHandler) GetCategories(w http.ResponseWriter, r *http.Request) {
	categories, err := h.service.GetCategories(r.Context())
	if err != nil {
		log.Printf("[ERROR] GetCategories failed: %v", err)
		common.WriteError(w, http.StatusInternalServerError, common.ErrInternalServer("failed to get categories"))
		return
	}

	common.WriteJSON(w, http.StatusOK, map[string]any{"categories": categories})
}

// --- Comments ---

// CreateComment handles creating a new comment on a listing.
func (h *MarketplaceHandler) CreateComment(w http.ResponseWriter, r *http.Request) {
	agent := middleware.GetAgent(r.Context())
	if agent == nil {
		common.WriteError(w, http.StatusUnauthorized, common.ErrUnauthorized("not authenticated"))
		return
	}

	listingIDStr := chi.URLParam(r, "id")
	listingID, err := uuid.Parse(listingIDStr)
	if err != nil {
		common.WriteError(w, http.StatusBadRequest, common.ErrBadRequest("invalid listing id"))
		return
	}

	var req marketplace.CreateCommentRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		common.WriteError(w, http.StatusBadRequest, common.ErrBadRequest("invalid request body"))
		return
	}

	comment, err := h.service.CreateComment(r.Context(), agent.ID, listingID, &req)
	if err != nil {
		common.WriteError(w, http.StatusBadRequest, common.ErrBadRequest(err.Error()))
		return
	}

	common.WriteJSON(w, http.StatusCreated, comment)
}

// GetListingComments handles getting comments for a listing.
func (h *MarketplaceHandler) GetListingComments(w http.ResponseWriter, r *http.Request) {
	listingIDStr := chi.URLParam(r, "id")
	listingID, err := uuid.Parse(listingIDStr)
	if err != nil {
		common.WriteError(w, http.StatusBadRequest, common.ErrBadRequest("invalid listing id"))
		return
	}

	limit := parseIntParam(r, "limit", 20)
	offset := parseIntParam(r, "offset", 0)

	result, err := h.service.GetListingComments(r.Context(), listingID, limit, offset)
	if err != nil {
		log.Printf("[ERROR] GetListingComments failed: %v (listingID=%s)", err, listingID)
		common.WriteError(w, http.StatusInternalServerError, common.ErrInternalServer("failed to get comments"))
		return
	}

	common.WriteJSON(w, http.StatusOK, result)
}

// GetCommentReplies handles getting replies to a comment.
func (h *MarketplaceHandler) GetCommentReplies(w http.ResponseWriter, r *http.Request) {
	commentIDStr := chi.URLParam(r, "commentId")
	commentID, err := uuid.Parse(commentIDStr)
	if err != nil {
		common.WriteError(w, http.StatusBadRequest, common.ErrBadRequest("invalid comment id"))
		return
	}

	replies, err := h.service.GetCommentReplies(r.Context(), commentID)
	if err != nil {
		log.Printf("[ERROR] GetCommentReplies failed: %v (commentID=%s)", err, commentID)
		common.WriteError(w, http.StatusInternalServerError, common.ErrInternalServer("failed to get replies"))
		return
	}

	common.WriteJSON(w, http.StatusOK, map[string]any{"replies": replies})
}

// DeleteComment handles deleting a comment.
func (h *MarketplaceHandler) DeleteComment(w http.ResponseWriter, r *http.Request) {
	agent := middleware.GetAgent(r.Context())
	if agent == nil {
		common.WriteError(w, http.StatusUnauthorized, common.ErrUnauthorized("not authenticated"))
		return
	}

	commentIDStr := chi.URLParam(r, "commentId")
	commentID, err := uuid.Parse(commentIDStr)
	if err != nil {
		common.WriteError(w, http.StatusBadRequest, common.ErrBadRequest("invalid comment id"))
		return
	}

	if err := h.service.DeleteComment(r.Context(), commentID, agent.ID); err != nil {
		common.WriteError(w, http.StatusBadRequest, common.ErrBadRequest(err.Error()))
		return
	}

	common.WriteJSON(w, http.StatusOK, map[string]string{"status": "deleted"})
}

// --- Request Comments ---

// CreateRequestComment handles creating a new comment on a request.
func (h *MarketplaceHandler) CreateRequestComment(w http.ResponseWriter, r *http.Request) {
	agent := middleware.GetAgent(r.Context())
	if agent == nil {
		common.WriteError(w, http.StatusUnauthorized, common.ErrUnauthorized("not authenticated"))
		return
	}

	requestIDStr := chi.URLParam(r, "id")
	requestID, err := uuid.Parse(requestIDStr)
	if err != nil {
		common.WriteError(w, http.StatusBadRequest, common.ErrBadRequest("invalid request id"))
		return
	}

	var req marketplace.CreateCommentRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		common.WriteError(w, http.StatusBadRequest, common.ErrBadRequest("invalid request body"))
		return
	}

	comment, err := h.service.CreateRequestComment(r.Context(), agent.ID, requestID, &req)
	if err != nil {
		common.WriteError(w, http.StatusBadRequest, common.ErrBadRequest(err.Error()))
		return
	}

	common.WriteJSON(w, http.StatusCreated, comment)
}

// GetRequestComments handles getting comments for a request.
func (h *MarketplaceHandler) GetRequestComments(w http.ResponseWriter, r *http.Request) {
	requestIDStr := chi.URLParam(r, "id")
	requestID, err := uuid.Parse(requestIDStr)
	if err != nil {
		common.WriteError(w, http.StatusBadRequest, common.ErrBadRequest("invalid request id"))
		return
	}

	limit := parseIntParam(r, "limit", 20)
	offset := parseIntParam(r, "offset", 0)

	result, err := h.service.GetRequestComments(r.Context(), requestID, limit, offset)
	if err != nil {
		log.Printf("[ERROR] GetRequestComments failed: %v (requestID=%s)", err, requestID)
		common.WriteError(w, http.StatusInternalServerError, common.ErrInternalServer("failed to get comments"))
		return
	}

	common.WriteJSON(w, http.StatusOK, result)
}

// GetRequestCommentReplies handles getting replies to a request comment.
func (h *MarketplaceHandler) GetRequestCommentReplies(w http.ResponseWriter, r *http.Request) {
	commentIDStr := chi.URLParam(r, "commentId")
	commentID, err := uuid.Parse(commentIDStr)
	if err != nil {
		common.WriteError(w, http.StatusBadRequest, common.ErrBadRequest("invalid comment id"))
		return
	}

	replies, err := h.service.GetRequestCommentReplies(r.Context(), commentID)
	if err != nil {
		log.Printf("[ERROR] GetRequestCommentReplies failed: %v (commentID=%s)", err, commentID)
		common.WriteError(w, http.StatusInternalServerError, common.ErrInternalServer("failed to get replies"))
		return
	}

	common.WriteJSON(w, http.StatusOK, map[string]any{"replies": replies})
}

// DeleteRequestComment handles deleting a request comment.
func (h *MarketplaceHandler) DeleteRequestComment(w http.ResponseWriter, r *http.Request) {
	agent := middleware.GetAgent(r.Context())
	if agent == nil {
		common.WriteError(w, http.StatusUnauthorized, common.ErrUnauthorized("not authenticated"))
		return
	}

	commentIDStr := chi.URLParam(r, "commentId")
	commentID, err := uuid.Parse(commentIDStr)
	if err != nil {
		common.WriteError(w, http.StatusBadRequest, common.ErrBadRequest("invalid comment id"))
		return
	}

	if err := h.service.DeleteRequestComment(r.Context(), commentID, agent.ID); err != nil {
		common.WriteError(w, http.StatusBadRequest, common.ErrBadRequest(err.Error()))
		return
	}

	common.WriteJSON(w, http.StatusOK, map[string]string{"status": "deleted"})
}

// --- Helpers ---

func parseIntParam(r *http.Request, name string, defaultVal int) int {
	valStr := r.URL.Query().Get(name)
	if valStr == "" {
		return defaultVal
	}
	val, err := strconv.Atoi(valStr)
	if err != nil {
		return defaultVal
	}
	return val
}